feat: add ConfirmationDialog and ErrorDialog components

+++ src/app/components/container/container.component.ts
@@ -1,19 +1,22 @@
 import { CommonModule } from '@angular/common';
 import { Component, OnInit } from '@angular/core';
 import { MatCardModule } from '@angular/material/card';
+import { MatDialog } from '@angular/material/dialog';
 import { MatPaginatorModule, PageEvent } from '@angular/material/paginator';
 import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { MatTabsModule } from '@angular/material/tabs';
 import { ActivatedRoute, Router } from '@angular/router';
-import { Observable, tap } from 'rxjs';
+import { catchError, Observable, of, tap } from 'rxjs';
+import { InventoryEventPage } from '../../model/inventory-event-page';
 import { Product } from '../../model/product';
 import { ProductPage } from '../../model/product-page';
+import { InventoryEventService } from '../../services/inventory-event.service';
 import { ProductService } from '../../services/product.service';
+import { ConfirmationDialogComponent } from '../confirmation-dialog/confirmation-dialog.component';
+import { ErrorDialogComponent } from '../error-dialog/error-dialog.component';
 import { InventoryEventTableComponent } from '../inventory-event-table/inventory-event-table.component';
 import { ProductTableComponent } from '../product-table/product-table.component';
-import { InventoryEventPage } from '../../model/inventory-event-page';
-import { InventoryEventService } from '../../services/inventory-event.service';
 
 @Component({
   selector: 'app-container',
@@ -38,6 +41,7 @@ export class ContainerComponent implements OnInit {
   productPageSize: number = 5;
 
   constructor(
+    private dialog: MatDialog,
     private inventoryEventService: InventoryEventService,
     private productService: ProductService,
     private router: Router,
@@ -59,6 +63,12 @@ export class ContainerComponent implements OnInit {
         tap(() => {
           (this.productPageIndex = pageEvent.pageIndex),
             (this.productPageSize = pageEvent.pageSize);
+        }),
+        catchError(() => {
+          this.dialog.open(ErrorDialogComponent, {
+            data: 'Erro ao carregar cursos',
+          })
+          return of({content: [], page: {totalElements: 0}} as ProductPage);
         })
       );
   }
@@ -72,13 +82,25 @@ export class ContainerComponent implements OnInit {
   }
 
   onDeleteProduct(product: Product): void {
-    this.productService.deleteProduct(product.id).subscribe({
-      next: () => {
-        this.loadProducts(),
-          this.snackBar.open('Produto excluído com sucesso', 'Ok', {
-            duration: 5000,
-          });
-      },
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      data: 'Você realmente deseja remover esse produto?',
+    });
+
+    dialogRef.afterClosed().subscribe((result: boolean) => {
+      if (result) {
+        this.productService.deleteProduct(product.id).subscribe({
+          next: () => {
+            this.loadProducts(),
+              this.snackBar.open('Produto excluído com sucesso', 'Ok', {
+                duration: 5000,
+              });
+          },
+          error: () =>
+            this.dialog.open(ErrorDialogComponent, {
+              data: 'Erro ao excluir produto',
+            }),
+        });
+      }
     });
   }
